+++ src/config/BoardList.js
@@ -1,0 +1,8 @@
+export default {
+  ConsoleBoard: { title: '콘솔 유저정보', params: {prefix: 'news', boardId: '1001' } },
+  PCBoard: { title: 'PC 정보 게시판', params: {prefix: 'news', boardId: '1003' } },
+  MobileBoard: { title: '모바일 정보게시판', params: {prefix: 'news', boardId: '1004' } },
+  BestBoard: { title: '많이 본 글/힛갤', params: {prefix: 'best/selection' } },
+  BestHumorBoard: { title: 'BEST - 일반유머', params: {prefix: 'best/humor' } },
+  BestHobbyBoard: { title: 'BEST - 취미 갤러리', params: {prefix: 'best/hobby' } },
+};
